- version up to episode 9:
https://www.youtube.com/watch?v=BFCYUR4zePc&t=500s

Added functionalities:
- game menu,
- movable pointer in the menu by "arrows" keys,
- "exit" menu button -> exiting from the game,
- "New Game" menu button -> start new game (game window),
- added background image in menu,
- fixed problem with loading images from Spritesheet (without pink and black color).

+++ src/core/core/gamestates/GS_Game.java
@@ -41,6 +41,6 @@ public class GS_Game extends GameState{
     public void render(Screen s){
         s.clear(0x0000ff);
 
-        s.renderSprite(x, y, this.s);
+        s.renderSprite((int)x, (int)y, this.s);
     }
 }
